broke up lexer.go into multiple files
